#132
Death check has been added to the game via a newly added hard-coded variable that is automatically added to the player repo called "Alive". At the end of each turn cycle it checks if the player is still alive and if not it triggers gameover. Gameover itself is not yet defined.