#100
Odi segments and all associated raw and variable based values are now stored as doubles instead of integers. This will allow multiplication or division based on more accurate values which should account for the need to calculate as percentages. Also, all things in the variable package had the "integer" references renamed to "number".