#66
Added the ability for default actions to be configured in the JSON in an "actions" section. This required a fair bit of generalization of some of the bits of AA code and a lot of file movement happened because of it. There are some pieces that are left half done that there are new tasks for and will be handled there.